requirements.txt
groq>=0.8.0
fastapi==0.111.1
uvicorn==0.30.1
python-dotenv==1.0.1
langdetect==1.0.9
google-generativeai
pypdf==4.3.1
pdfplumber
llama-index>=0.11.0
llama-index-llms-gemini
llama-index-embeddings-huggingface
llama-index-retrievers-bm25
rank-bm25
llama-index-postprocessor-flag-embedding-reranker
spacy
pyspellchecker
sentence-transformers
